requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
bcrypt==4.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
idna==3.10
mysql-connector-python==9.0.0
passlib==1.7.4
pydantic==2.9.2
pydantic_core==2.23.4
redis==5.0.8
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6